Refactor ministeries with shared base class
